jobs/kpi-forecasting/requirements.txt
google-cloud-bigquery==1.27.2
google-cloud-bigquery-storage==1.0.0
google-cloud-storage==1.31.0
Jinja2==2.11.2
matplotlib==3.3.2
numpy==1.21.4
pandas==1.3.5
pytest==6.0.2
pytest-black==0.3.11
pytest-flake8==1.0.6
plotly==4.9.0
pyyaml==6.0
holidays==0.13
prophet==1.0.1
pyarrow==7.0
